move common input types into separate files
