+++ src/phoenix/server/api/input_types/TemplateOptions.py
@@ -1,0 +1,10 @@
+import strawberry
+from strawberry.scalars import JSON
+
+from phoenix.server.api.types.TemplateLanguage import TemplateLanguage
+
+
+@strawberry.input
+class TemplateOptions:
+    variables: JSON
+    language: TemplateLanguage
